requirements.txt
torch
numpy
PyYAML
transformers
fugashi
unidic-lite
ipadic
tqdm